prepare encapsulated based nvim

+++ exe/nide.lua
@@ -1,13 +1,24 @@
 local doExecute = true -- false for debugging
 
 -- Funktion, um den Pfad der aktuellen Lua-Datei zu ermitteln
-local function getCurrentFilePath()
+local function get_current_file_path()
     local info = debug.getinfo(1, "S")
     return info.source:match("@(.*)$")
 end
 
+local function get_nvim_count()
+    local handle = io.popen("pgrep -x -c nvim 2>/dev/null")
+    if not handle then return 0 end
+
+    local result = handle:read("*a")
+    handle:close()
+
+    local count = tonumber(result:match("%d+"))
+    return count or 0
+end
+
 -- Pfad der aktuellen Datei ermitteln
-local currentFilePath = getCurrentFilePath()
+local currentFilePath = get_current_file_path()
 
 -- Pfad zum 'python'-Verzeichnis ermitteln (relativ zum aktuellen Skript)
 local localPath = currentFilePath:sub(1, currentFilePath:find("/[^/]*$") - 1)
@@ -26,10 +37,8 @@ local in_tmux = os.getenv("TMUX") ~= nil
 
 if (doExecute and not in_tmux) then os.execute("tmux new-session -d -s ide") end
 
--- fixme: grep current user id
 local runtime_dir = os.getenv("XDG_RUNTIME_DIR")
 if not runtime_dir or runtime_dir == '' then runtime_dir = "/tmp" end
-
 local runtime_nvim = runtime_dir .. '/nvim'
 
 local attr = lfs.attributes(runtime_nvim)
@@ -43,16 +52,18 @@ local pipepath = nil
 local cwd = lfs.currentdir()
 local cmd_front = 'tmux send-keys -t ide C-m \"'
 local cmd_mid = nil
+local cmd_tail = nil
 if cwd:find("encapsulated") then
     pipepath = runtime_nvim .. '/ide-enc.pipe'
     cmd_mid = conf.config.docker_path .. ' run -it ' ..
         conf.eenv .. ' ' .. rundocker.map(runtime_nvim)  ..
         conf.config.image
+    cmd_tail = ' \\"' .. conf.config.nvim_path .. ' --listen ' .. pipepath .. '\\"\" C-m'
 else
     pipepath = runtime_nvim .. '/ide.pipe'
     cmd_mid = conf.env
+    cmd_tail = ' ' .. conf.config.nvim_path .. ' --listen ' .. pipepath .. '\" C-m'
 end
-local cmd_tail = ' ' .. conf.config.nvim_path .. ' --listen ' .. pipepath .. '\" C-m'
 local cmd = cmd_front .. cmd_mid .. cmd_tail
 print(cmd)
 
@@ -68,5 +79,7 @@ end
 
 local pdir = require("pl.dir")
 local nvim_state = home .. "/.local/state/nvim"
-print("delete: " .. nvim_state)
-pdir.rmtree(nvim_state)
+if get_nvim_count() == 0 then
+    print("delete: " .. nvim_state)
+    pdir.rmtree(nvim_state)
+end
